+++ docs/whats_new.rst
@@ -41,6 +41,7 @@ API changes
 ^^^^^^^^^^^
 
 - Added :func:`mne_hfo.io.events_to_annotations` to go from ``*events.tsv`` to ``*annotations.tsv`` files, by `Adam Li`_ (:gh:`10`)
+- Added :func:`mne_hfo.sklearn.make_Xy_sklearn` to format data into scikit-learn compatible data structures for the sake of running hyper-parameter searches with ``SearchCV`` functions, by `Adam Li`_ (:gh:`15`)
 
 Requirements
 ^^^^^^^^^^^^
