+++ examples/detect_hfos.py
@@ -107,7 +107,7 @@ detector = RMSDetector(**kwargs)
 detector.fit(X=raw)
 
 # get the event dataframe
-event_df = detector.hfo_df
+event_df = detector.hfo_event_df
 print(event_df.head())
 
 ###############################################################################
