+++ tests/test_score.py
@@ -1,7 +1,7 @@
 from mne_hfo import create_annotations_df
-from mne_hfo.scores import accuracy, true_positive_rate, \
+from mne_hfo.score import accuracy, true_positive_rate, \
     false_negative_rate, false_discovery_rate, precision
-from mne_hfo.utils import _make_ydf_sklearn
+from mne_hfo.sklearn import _make_ydf_sklearn
 
 
 def test_match_detection_scoring_df():
